requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
alembic>=1.7.0
pytest>=6.2.5
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
psycopg2-binary>=2.9.1
httpx>=0.19.0
bcrypt>=3.2.0
python-dotenv>=0.19.0
email-validator>=1.1.0
asyncpg>=0.23.0
pyotp>=2.8.0
qrcode>=7.3.1
pillow>=9.0.0